// EcomWeb/src/api/client/models/ProductImage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Product } from './Product';
export type ProductImage = {
    isDeleted?: boolean;
    readonly id?: number;
    readonly imageUrl?: string | null;
    readonly isPrimary?: boolean;
    readonly productId?: number;
    product?: Product;
};